# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TetraCore Hub API
  description: |
    TetraCore Hub is a distributed task orchestration and management system.

    ## Authentication
    The API supports two authentication methods:
    - **JWT Token**: Use `Authorization: Bearer <token>` header
    - **API Key**: Use `X-API-Key: <key>` header or `Authorization: Bearer tk_<key>`

    ## Rate Limiting
    Default rate limit is 1000 requests per minute. API keys may have custom limits.
  version: 2.0.0
  contact:
    name: TetraCore Support
    email: support@tetra-core.website
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://hub.tetra-core.website
    description: Production
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    CreateUserRequest:
      type: object
      required: [username, password, role_id]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role_id:
          type: string
...